refactor: rename tool_handler.py to tool_state.py
